fix(demo-be-rsax): fix logout to use Authorization header and quantity column type
Logout now extracts token from Authorization header (matching e2e tests)
instead of request body. Changed quantity column from REAL to DOUBLE
PRECISION for PostgreSQL f64 compatibility.

+++ apps/demo-be-rsax/src/db/migrations/003_expenses.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS expenses (
     description TEXT NOT NULL,
     date TEXT NOT NULL,
     entry_type TEXT NOT NULL,
-    quantity REAL,
+    quantity DOUBLE PRECISION,
     unit TEXT,
     created_at TEXT NOT NULL,
     updated_at TEXT NOT NULL,
